requirements.txt
asgiref==3.8.1
Babel==2.17.0
Django==5.2.2
django-extra-views==0.15.0
django-haystack==3.3.0
django-oscar==4.0
django-phonenumber-field==8.1.0
django-tables2==2.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory_boy==3.3.3
Faker==37.3.0
gunicorn==23.0.0
packaging==25.0
phonenumbers==9.0.6
pillow==11.2.1
purl==1.6
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary>=2.9
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
